requirements.txt
playwright
groq
python-dotenv
